fix: top-up queue limit to uint8

+++ script/DeployBase.s.sol
@@ -72,7 +72,7 @@ struct DeployParams {
     uint256 stakingModuleId;
     bytes32 moduleType;
     address generalDelayedPenaltyReporter;
-    uint32 topUpQueueLimit;
+    uint8 topUpQueueLimit;
     // ParametersRegistry
     uint256 queueLowestPriority;
     uint256 defaultKeyRemovalCharge;
